api/MalariaCellsClassification/flask/requirements.txt
Flask==1.1.1
Keras==2.3.1
tensorflow==2.0.0
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
gunicorn==20.0.4